refactor some
